Use chembl_downloader for generating library
The chembl_downloader automates downloading the chembl SDF dump, and there's a little code that wraps gunzipping it and creating a ForwardSDMolSupplier in a context generator. This will hopefully make it a bit easier for someone else to use this code!

I wrote more about it on my blog https://cthoyt.com/2021/08/04/taming-chembl-sdf.html